user: Refactor class-interface name
Description:
refactored user interactor

Expected behavior:
N/A
