--- src/main/java/com/splitscale/fordastore/core/user/InteractableUser.java
@@ -1,12 +1,0 @@
-package com.splitscale.fordastore.core.user;
-
-import com.splitscale.fordastore.core.auth.TokenizedBasicAuth;
-import com.splitscale.fordastore.core.http.Message;
-
-public interface InteractableUser {
-  DisplayableUser authUser(TokenizedBasicAuth request);
-
-  Message deleteUser(TokenizedBasicAuth request, String uid);
-
-  Message updateUser(TokenizedBasicAuth request, String uid);
-}
